requirements.txt
annotated-types==0.7.0
asyncio==3.4.3
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
dotenv==0.9.9
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.163.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.1
grpcio==1.71.0rc2
grpcio-status==1.71.0rc2
httplib2==0.22.0
idna==3.10
joblib==1.4.2
numpy==2.2.3
pdfminer.six==20231228
pdfplumber==0.11.5
pillow==11.1.0
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
pypdfium2==4.30.1
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
scikit-learn==1.6.1
scipy==1.15.2
threadpoolctl==3.5.0
tqdm==4.67.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
